Moved PriorityQueue to the Concurrent::Collection module.
